requirements.txt
django==3.0.8
djangorestframework
markdown
django-filter
django-mysql==3.8.1
mysqlclient
